feat(sharedServiceForCreateAccountAndAddWorkplace): Create new-workplace-not-found directive and get create account component to extend it
