+++ src/app/shared/directives/create-workplace/new-workplace-not-found/new-workplace-not-found.directive.ts
@@ -1,0 +1,102 @@
+import { AfterViewInit, Directive, ElementRef, OnInit, ViewChild } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { ActivatedRoute, Router } from '@angular/router';
+import { ErrorDetails } from '@core/model/errorSummary.model';
+import { Establishment } from '@core/model/establishment.model';
+import { BackService } from '@core/services/back.service';
+import { ErrorSummaryService } from '@core/services/error-summary.service';
+import { EstablishmentService } from '@core/services/establishment.service';
+import { RegistrationService } from '@core/services/registration.service';
+import { SanitizePostcodeUtil } from '@core/utils/sanitize-postcode-util';
+
+@Directive()
+export class NewWorkplaceNotFoundDirective implements OnInit, AfterViewInit {
+  @ViewChild('formEl') formEl: ElementRef;
+  public form: FormGroup;
+  public formErrorsMap: Array<ErrorDetails>;
+  public serverError: string;
+  public submitted: boolean;
+  public workplace: Establishment;
+  public isParent: boolean;
+  public postcodeOrLocationId: string;
+  protected searchMethod: string;
+  protected flow: string;
+
+  constructor(
+    protected establishmentService: EstablishmentService,
+    protected formBuilder: FormBuilder,
+    public backService: BackService,
+    protected errorSummaryService: ErrorSummaryService,
+    protected registrationService: RegistrationService,
+    protected router: Router,
+    protected route: ActivatedRoute,
+  ) {}
+
+  ngOnInit(): void {
+    this.flow = this.route.snapshot.parent.url[0].path;
+    this.workplace = this.establishmentService.primaryWorkplace;
+    this.isParent = this.workplace?.isParent ? true : false;
+    this.sanitizePostcode();
+    this.setBackLink();
+    this.setupForm();
+    this.setupFormErrorsMap();
+  }
+
+  ngAfterViewInit(): void {
+    this.errorSummaryService.formEl$.next(this.formEl);
+  }
+
+  public sanitizePostcode(): void {
+    this.postcodeOrLocationId = this.registrationService.postcodeOrLocationId$.value;
+    this.searchMethod = this.registrationService.searchMethod$.value;
+
+    if (this.searchMethod === 'postcode') {
+      this.postcodeOrLocationId = SanitizePostcodeUtil.sanitizePostcode(this.postcodeOrLocationId);
+    }
+  }
+
+  private setupForm(): void {
+    this.form = this.formBuilder.group({
+      useDifferentLocationIdOrPostcode: [null, { validators: Validators.required, updateOn: 'submit' }],
+    });
+  }
+
+  private setupFormErrorsMap(): void {
+    this.formErrorsMap = [
+      {
+        item: 'useDifferentLocationIdOrPostcode',
+        type: [
+          {
+            name: 'required',
+            message: `Select yes if you want to try a different location ID or postcode`,
+          },
+        ],
+      },
+    ];
+  }
+
+  public setBackLink(): void {
+    this.backService.setBackLink({ url: [this.flow, 'find-workplace'] });
+  }
+
+  public getErrorMessage(item: string): string {
+    const errorType = Object.keys(this.form.get(item).errors)[0];
+    return this.errorSummaryService.getFormErrorMessage(item, errorType, this.formErrorsMap);
+  }
+
+  public onSubmit(): void {
+    this.submitted = true;
+
+    if (this.form.valid) {
+      const useDifferentLocationIdOrPostcode = this.form.get('useDifferentLocationIdOrPostcode');
+      this.registrationService.workplaceNotFound$.next(true);
+      if (useDifferentLocationIdOrPostcode.value === 'yes') {
+        this.router.navigate([`/${this.flow}`, 'find-workplace']);
+      } else {
+        this.router.navigate([`/${this.flow}`, 'workplace-name-address']);
+      }
+    } else {
+      this.errorSummaryService.scrollToErrorSummary();
+    }
+  }
+}
